refactor: move remaining and overtime calculation to time utils
