+++ src/app/services/time-tracking/time.service.ts
@@ -3,7 +3,7 @@ import { TimeTable } from './time.table';
 import { SettingsTable } from '../settings/settings.table';
 import { combineLatest, map, shareReplay } from 'rxjs';
 import { Duration } from 'luxon';
-import { multiplyDuration } from '../time.utils';
+import { calculateRemainingAndOvertime, multiplyDuration } from '../time.utils';
 
 @Injectable({ providedIn: 'root' })
 export class TimeService {
@@ -11,23 +11,9 @@ export class TimeService {
     private readonly settingsTable = inject(SettingsTable);
     readonly today$ = combineLatest([this.timeTable.todayGroup$(), this.settingsTable.current$()]).pipe(
         map(([today, settings]) => {
-            console.time('calculation');
             const workPerDay = Duration.fromMillis(settings.workPerDay);
 
-            const { duration } = today;
-            const durationMilliseconds = duration.toMillis();
-
-            let remainingTime: Duration | undefined;
-            let overtime: Duration | undefined;
-
-            if (durationMilliseconds > settings.workPerDay) {
-                overtime = duration.minus(workPerDay);
-            }
-
-            if (durationMilliseconds < settings.workPerDay) {
-                remainingTime = workPerDay.minus(duration);
-            }
-            console.timeEnd('calculation');
+            const { remainingTime, overtime } = calculateRemainingAndOvertime(workPerDay, today.duration);
 
             return {
                 ...today,
@@ -45,19 +31,7 @@ export class TimeService {
             const nominalTime = multiplyDuration(workPerDay, amountOfDaysWorked);
             const actualTime = groups.reduce((sum, current) => sum.plus(current.duration), Duration.fromMillis(0));
 
-            const nominalTimeMilliseconds = nominalTime.toMillis();
-            const actualTimeMilliseconds = actualTime.toMillis();
-
-            let remainingTime: Duration | undefined;
-            let overtime: Duration | undefined;
-
-            if (nominalTimeMilliseconds > actualTimeMilliseconds) {
-                remainingTime = nominalTime.minus(actualTime);
-            }
-
-            if (nominalTimeMilliseconds < actualTimeMilliseconds) {
-                overtime = actualTime.minus(nominalTime);
-            }
+            const { remainingTime, overtime } = calculateRemainingAndOvertime(nominalTime, actualTime);
 
             return {
                 nominalTime,
